There seem to be timing problems with deleting the directories. I've only seen this on Windows, but I have no idea how hardware may affect it. (My Linux and Windows systems are the same box, but they OS's don't manage memory the same way at all, so I don't really know what the problem is.)
So let's just not delete the database directories when we exit. We have
always been deleting them on startup anyway, so they should always be
clean.


git-svn-id: https://svn.internet2.edu/svn/i2mi/trunk@4972 bc24b658-cf39-4249-b827-e3526eaaeb42

+++ ldappc/src/test/java/edu/internet2/middleware/ldappcTest/wrappers/DatabaseWrapperTestSetup.java
@@ -113,19 +113,19 @@ public class DatabaseWrapperTestSetup extends TestSetup {
             }
         }
 
-        // Delete the grouper database.
-        if (useEmbeddedGrouper) {
-            File dbDir = new File(grouperDbTarget);
-            FileUtils.deleteDirectory(dbDir);
-        }
-
-        // Delete the signet database.
-        if (useEmbeddedSignet) {
-            File dbDir = new File(signetDbTarget);
-            FileUtils.deleteDirectory(dbDir);
-        }
-
-        super.tearDown();
+        if (false) {
+			// Delete the grouper database.
+			if (useEmbeddedGrouper) {
+				File dbDir = new File(grouperDbTarget);
+				FileUtils.deleteDirectory(dbDir);
+			}
+			// Delete the signet database.
+			if (useEmbeddedSignet) {
+				File dbDir = new File(signetDbTarget);
+				FileUtils.deleteDirectory(dbDir);
+			}
+		}
+		super.tearDown();
     }
 
     /**
